Make the propagator configurable
